requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
dj-email-url==1.0.1
Django==3.1.5
django-cache-url==3.1.2
environs==8.0.0
future==0.18.2
marshmallow==3.10.0
numpy==1.19.3
Pillow==7.2.0
python-dotenv==0.15.0
pytz==2020.5
scipy==1.6.0
sqlparse==0.4.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.6.0+cpu
torchvision==0.7.0+cpu
